pyproject.toml
[tool.poetry]
name = "from-jupyter-to-deploy"
version = "0.1.0"
description = ""
authors = ["guimarczewski <guimarczewski@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.3"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
taskipy = "^1.12.2"
pytest = "^7.4.4"
httpx = "^0.26.0"
pydantic = "^2.5.3"
openpyxl = "^3.1.2"
pandas = "^2.2.0"
plotly = "^5.18.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.25"
streamlit = "^1.30.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
run = "uvicorn api.app.main:app --reload"
test = "pytest -v api/tests/tests.py"
kill = "kill -9 $(lsof -t -i:8000)"
